Frontend: Changed all html classes to use class-name style
Frontend: Added couple isRequired prop checkers

Frontend: Added usage instructions for restart-docker-dev-container.sh

+++ restart-docker-dev-container.sh
@@ -10,5 +10,10 @@ if [ "$1" == "backend" ]; then
     docker-compose -f docker-compose.dev.yml -p ekanbandev rm -f ekanban-mongodb
 fi
 
+if [ -z "$1" ]; then
+    echo -e "\n \033[1mUsage: ./restart-docker-dev-container.sh container-name (frontend or backend)\033[m\n"
+    exit
+fi
+
 docker-compose -f docker-compose.dev.yml -p ekanbandev rm -f ekanban-$1
 docker-compose -f docker-compose.dev.yml -p ekanbandev up -d --build --force-recreate ekanban-$1
